+++ shop/tests/shipping.py
@@ -180,7 +180,7 @@ class FlatRateShippingTestCase(TestCase):
 
         # User logged in (no order)
         view = self.backend.view_process_order(self.request)
-        self.assertEqual(view.get('location', None), '/')
+        self.assertEqual(view.get('location', None), '/shop/cart/')
 
         # User logged in with order
         order = Order()